+++ packages/coding-agent/test/tiny-worker-env.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "bun:test";
 import * as path from "node:path";
 import { nativeLibraryPathOverlay } from "@oh-my-pi/pi-coding-agent/subprocess/worker-client";
-import { tinyWorkerEnvOverlay } from "@oh-my-pi/pi-coding-agent/tiny/title-client";
+import { tinyWorkerEnvOverlay } from "@oh-my-pi/pi-coding-agent/tiny/model-client";
 import { tinyWorkerEndpoint, tinyWorkerLogPath } from "@oh-my-pi/pi-coding-agent/tiny/title-protocol";
 
 describe("tinyWorkerEnvOverlay", () => {